Move interface to ghcrateengine and implement function in the ghcratengine instead of fuel price storer
